requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==4.2.17
django-bootstrap-icons==0.8.7
django-bootstrap5==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
idna==3.10
pillow==11.0.0
PyJWT==1.7.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
